Fix menu width for owner-drawn items using bitmap in wxMSW
Do _not_ provide a valid HBITMAP for them, as this seems to break
measuring such items by not using our WM_MEASUREITEM handler at all.

Note that we could also use HBMMENU_CALLBACK for the bitmap, which also
fixes the problem, but it seems better to just not do anything at all.

Closes #23230.

Co-Authored-by: Jacob Gillespie <jacobgillespie@gmail.com>